added ScraperError for debugging, improved debugging

+++ app/jobs/remove_duplicates_job.rb
@@ -7,8 +7,12 @@ class RemoveDuplicatesJob < ApplicationJob
     @jobs.each do |job|
       begin
         @duplicates = Job.all.where(title: job.title).where(company: job.company).where(salary: job.salary).where(search_id: job.search_id)
-        @duplicates[(1..-1)].each do |dup|
-          dup.destroy
+        if @dublicates.nil?
+          puts "#{@duplicates.first.title} has no duplicates"
+        else
+          @duplicates[(1..-1)].each do |dup|
+            dup.destroy
+          end
         end
       rescue StandardError => e
         puts e.message
